Style: 폴더/파일  구조 변경, 코드 분리
