+++ src/app/App.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Routes, Route, useNavigate } from 'react-router-dom';
 
-import Header from '../components/Header';
+import Header from '../components/header/Header';
 import Main from '../pages/Main';
 import Login from '../pages/Login';
 import StudyPage from '../pages/StudyPage';
