+++ src/util/tensorflow/posepoints.js
@@ -1,0 +1,11 @@
+// const head = pose.keypoints[0].position;
+// const left_Eye = pose.keypoints[1].position;
+// const right_Eye = pose.keypoints[2].position;
+// const left_leftEar = pose.keypoints[3].position;
+// const right_left = pose.keypoints[4].position;
+// const left_Shoulder = pose.keypoints[5].position;
+// const right_Shoulder = pose.keypoints[6].position;
+// const left_Elbow = pose.keypoints[7].position;
+// const right_Elbow = pose.keypoints[8].position;
+// const left_Wrist = pose.keypoints[9].position;
+// const right_Wrist = pose.keypoints[10].position;
